Added flexible aria-label to <nav> element; removed unnecessary dependencies and rearranged file locations wtihin projects.
