--- ThePensionsRegulator.Frontend.Umbraco/Services/ITprSideNavigationLinksService.cs
@@ -1,10 +1,0 @@
-﻿using ThePensionsRegulator.Frontend.Umbraco.Models;
-using Umbraco.Cms.Core.Models.PublishedContent;
-
-namespace ThePensionsRegulator.Frontend.Umbraco.Services
-{
-    public interface ITprSideNavigationLinksService
-    {
-        public TprSideNavigationViewModel GetLinks(IPublishedContent currentPage);
-    }
-}
