+++ ThePensionsRegulator.Frontend/Services/ITprSideNavigationLinksService.cs
@@ -1,0 +1,9 @@
+﻿using ThePensionsRegulator.Frontend.Models;
+
+namespace ThePensionsRegulator.Frontend.Services
+{
+    public interface ITprSideNavigationLinksService
+    {
+        public TprSideNavigationViewModel GetLinks();
+    }
+}
